# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HospitalCreate": {
        "additionalProperties": false,
        "description": "Payload to register a new hospital (Admin only).",
        "properties": {
          "address": {
            "maxLength": 500,
            "minLength": 5,
            "title": "Address",
            "type": "string"
          },
          "capabilities": {
            "description": "Registered facilities/specialties, e.g. ['ICU', 'TRAUMA_LEVEL_1']",
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "contact_number": {
            "maxLength": 50,
            "minLength": 5,
            "title": "Contact Number",
            "type": "string"
          },
          "latitude": {
            "description": "Latitude between -90 and 90",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "description": "Longitude between -180 and 180",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "registration_identifier": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Registration Identifier",
            "type": "string"
          },
          "reported_availability": {
            "additionalProperties": true,
            "description": "Initial reported capacity, e.g. {'icu_beds': 5, 'accepting_patients': true}",
            "title": "Reported Availability",
            "type": "object"
          }
        },
        "required": [
          "name",
          "registration_identifier",
          "address",
          "latitude",
          "longitude",
          "contact_number"
        ],
        "title": "HospitalCreate",
        "type": "object"
      },
...
      "HospitalRequestDeclinePayload": {
        "additionalProperties": false,
        "description": "Payload when a hospital declines an admission request.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional clinical or operational reason for declining",
            "title": "Reason"
          }
        },
        "title": "HospitalRequestDeclinePayload",
        "type": "object"
      },
      "HospitalRequestResponse": {
        "description": "Public representation of an emergency admission request to a hospital.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "emergency_id": {
            "format": "uuid",
            "title": "Emergency Id",
            "type": "string"
          },
          "hospital_id": {
            "format": "uuid",
            "title": "Hospital Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "responded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At"
          },
          "response_deadline": {
            "format": "date-time",
            "title": "Response Deadline",
            "type": "string"
          },
          "response_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Reason"
          },
          "status": {
            "$ref": "#/components/schemas/HospitalRequestStatus"
          }
        },
        "required": [
          "id",
          "emergency_id",
          "hospital_id",
          "status",
          "created_at",
          "response_deadline"
        ],
        "title": "HospitalRequestResponse",
        "type": "object"
      },
      "HospitalRequestStatus": {
        "description": "Status of a hospital dispatch/admission request.",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "DECLINED",
          "EXPIRED",
          "CANCELLED"
        ],
        "title": "HospitalRequestStatus",
        "type": "string"
      },
      "HospitalResponse": {
        "description": "Public hospital representation.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "availability_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Updated At"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "contact_number": {
            "title": "Contact Number",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "registration_identifier": {
            "title": "Registration Identifier",
            "type": "string"
          },
          "reported_availability": {
            "additionalProperties": true,
            "title": "Reported Availability",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "registration_identifier",
          "address",
          "latitude",
          "longitude",
          "contact_number",
          "capabilities",
          "reported_availability",
          "created_at",
          "updated_at"
        ],
        "title": "HospitalResponse",
        "type": "object"
      },
...
      "PaginatedResponse_HospitalResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HospitalResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[HospitalResponse]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the Emergency Response Coordination System.",
    "title": "Emergency Response Coordination System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/emergencies/{emergency_id}/requests": {
      "get": {
        "description": "Retrieve all hospital requests generated for this emergency incident.",
        "operationId": "get_emergency_requests_api_v1_emergencies__emergency_id__requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "emergency_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Emergency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HospitalRequestResponse"
                  },
                  "title": "Response Get Emergency Requests Api V1 Emergencies  Emergency Id  Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List admission requests dispatched for an emergency",
        "tags": [
          "Emergencies"
        ]
      }
    },
...
    "/api/v1/hospital-requests": {
      "get": {
        "description": "Retrieve hospital admission requests.\n\nHospital staff only see requests directed to their assigned hospital.\nAdministrators can see all requests.",
        "operationId": "list_hospital_requests_api_v1_hospital_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/HospitalRequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HospitalRequestResponse"
                  },
                  "title": "Response List Hospital Requests Api V1 Hospital Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List admission requests for current hospital",
        "tags": [
          "Hospital Requests"
        ]
      }
    },
    "/api/v1/hospital-requests/{request_id}": {
      "get": {
        "description": "Retrieve details of a specific hospital request.",
        "operationId": "get_hospital_request_api_v1_hospital_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get hospital admission request details",
        "tags": [
          "Hospital Requests"
        ]
      }
    },
    "/api/v1/hospital-requests/{request_id}/accept": {
      "post": {
        "description": "Accept an emergency request with strict PostgreSQL row-level concurrency locking.\n\nGuarantees that competing concurrent acceptance attempts from multiple hospitals\nwill safely result in exactly one winning destination assignment and a 409 Conflict\nfor losing requests.",
        "operationId": "accept_hospital_request_api_v1_hospital_requests__request_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept an inbound emergency admission request (Concurrency Safe)",
        "tags": [
          "Hospital Requests"
        ]
      }
    },
    "/api/v1/hospital-requests/{request_id}/decline": {
      "post": {
        "description": "Decline an admission request. Idempotent if already declined.",
        "operationId": "decline_hospital_request_api_v1_hospital_requests__request_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HospitalRequestDeclinePayload"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decline an inbound emergency admission request",
        "tags": [
          "Hospital Requests"
        ]
      }
    },
    "/api/v1/hospitals": {
      "get": {
        "description": "Retrieve a paginated directory of registered hospitals.\n\nAccessible to any authenticated user.",
        "operationId": "list_hospitals_api_v1_hospitals_get",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_HospitalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List registered hospitals",
        "tags": [
          "Hospitals"
        ]
      },
      "post": {
        "description": "Create and register a hospital in the system.\n\nRestricted to ADMIN users.",
        "operationId": "create_hospital_api_v1_hospitals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HospitalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register a new hospital (Admin only)",
        "tags": [
          "Hospitals"
        ]
      }
    },
...
  }
}